refactor 랜덤 클래스 -> 랜덤 요소 의존성 제거를 위하여 인터페이스, 구현체 로 분리
