+++ src/main/java/racing/service/GameRule.java
@@ -1,0 +1,8 @@
+package racing.service;
+
+public interface GameRule {
+    public int getPossibleInt();
+
+    public int makeRandomInt();
+
+}
