+++ src/main/java/racing/service/RandomGameRule.java
@@ -1,17 +1,18 @@
-package racing.util;
+package racing.service;
 
 import java.util.Random;
 
-public class RandomInt {
+public class RandomGameRule implements GameRule {
 
     private static final int RANDOM_INT_MAX = 10;
     private static final int POSSIBLE_INT = 4;
 
-    private static int makeRandomInt() {
+    public int makeRandomInt() {
         Random random = new Random();
         return random.nextInt(RANDOM_INT_MAX);
     }
-    public static int getPossibleInt() {
+
+    public int getPossibleInt() {
         return (POSSIBLE_INT <= makeRandomInt()) ? 1 : 0;
     }
 }
