+++ src/test/java/racing/domain/CarTest.java
@@ -2,10 +2,8 @@ package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racing.domain.Car;
-import racing.util.RandomInt;
-
-import java.util.Random;
+import racing.service.GameRule;
+import racing.service.RandomGameRule;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -15,22 +13,25 @@ public class CarTest {
     @DisplayName("0~8 랜덤 int 생성 ")
     void randomInt() {
 
-        int randomInt = RandomInt.getPossibleInt();
-        assertThat(randomInt).isLessThan(9);
+        GameRule gameRule = new RandomGameRule();
+        int randomNumber = gameRule.getPossibleInt();
+        assertThat(randomNumber).isLessThan(9);
     }
 
     @Test
     @DisplayName("자동차 이동 시 랜덤 int 만큼 거리 변화")
     void customDelimiterSum() {
 
+        GameRule gameRule = new RandomGameRule();
+
         Car car = new Car(0);
-        int firstRandomInt = RandomInt.getPossibleInt();
+        int firstRandomInt = gameRule.getPossibleInt();
         Car firstMovedCar = car.move(firstRandomInt);
 
-        int secondRandomInt = RandomInt.getPossibleInt();
+        int secondRandomInt = gameRule.getPossibleInt();
         Car secondMovedCar = car.move(secondRandomInt);
 
-        assertThat(firstMovedCar.getLocation()).isEqualTo(firstRandomInt);
-        assertThat(secondMovedCar.getLocation() - firstMovedCar.getLocation()).isEqualTo(secondRandomInt);
+        assertThat(firstMovedCar.location()).isEqualTo(firstRandomInt);
+        assertThat(secondMovedCar.location() - firstMovedCar.location()).isEqualTo(secondRandomInt);
     }
 }
